requirements.txt
mysql-connector-python==9.1.0
scapy=2.6.1
